Update app creation to take advantage of DF Proxy
We no longer expose ports for apps. Instead, we join them to DockerFlow
Proxy's proxy network, and set appropriate labels so Proxy knows to
keep track of the app's web service.

Getting the DF Proxy service started and configured is part of the inital
PaaS deployment. And so is  creating appropriate DNS A records to point
at your manager node.