Add config model
We can use this to store configuration data about the PaaS. At the moment,
I need a place to store the domain name associated with the PaaS. Rows
in this table have a key and a value, each row represents a configuration
setting. This would be more appropriately handled by a key value store, but
this will do for now.