Change app field name
I had initially created a url field for an app where I was imagining storing the
complete URL to the app. But that is a bad idea since we already store the main
domain in the Config table.

Instead, we'll only save the app's subdomain. Anyone needing the complete
URL can combine the app's subdomain with the domain found in Configs.