Switch back to using url field
I thought the advantage of not using url, but just subdomain, would be that
it would make it easy to move the PaaS to a different domain without having
to update all the app urls. But we'd have to do that anyway since the complete
url is stored in DockerFlow Proxy.

Since the advantage of normalizing here is already spoiled, we might as well
make the app model easier to use and just store the complete url.