Fix layout bug differently
We previously fixed the main section overflowing the body by setting an
overflow-y property, but I couldn't understand why this worked. Now we
fix it the issue directly: we shouldn't have been setting a height on body,
but only a min-height. Now it grows with the content, but is still at least
a full viewport high.