Fix exec command to connect to app database
Before, we were running exec commands inside of a one-off container, but
we weren't doing the things we needed to allow the container to connect to
the app's database (for `db:migrate` or whatever).

First, we set the appropriate environment variables, so the app knows the db
connection info. Second we join the app's network.

The second step isn't as striaghtforward as it seems: by default, standalone containers cannot connect to Swarm overlay networks, only Swarm services can. To get around this, we change the way we create networks for apps: now, we set them to be encrypted and we make them attachable.