Emit events selectively during image build
If we get an error, we want to emit a special ERROR event (that will eventually)
cause the server to stop streaming to the client, and the client to realize the
build failed. If we hit a "Succesfully tagged..." message, we know we're done, and
so resolve the promise. Otherwise, it's just a normal intermediate step and we
emit it.