Add health-check function to DockerWrapper
This function gets all the tasks associated with the current app's
service and checks for failed tasks. If the number of tasks that
have failed in the last minute exceeds some threshhold, that means
your app is crashing and being re-started very rapidly, indicating
a problem.

For now, the threshhold I've selected is 3. This is because the
Docker API doesn't seem to send back more than 5 tasks for a given
service. Even for an app that can't boot at all, there will typically
be 1 task that reports as 'ready' (which hasn't yet had a chance to
crash).

Another health-check we should add is to send a simple curl request
to the application's root and make sure we get back a non-4/500
response (Beanstalk does this).