Add websocket server for handling app health checks
We're going to have the app show page establish a websocket connection
with a websocket server. This server will repeatedly check the
health of the app every second and send the result to the client.
This stops whenever the client leaves the show page.

We're using the node 'ws' library for the websocket server. Rather
than make the websocket server listen on a port, we let our main
app server received the websocket upgrade request, and when it
does, it just passes this on to the websocket server.

It's not ideal having application logic like this in our bin/www
file. We need to get it out, and just import a function that we
can pass to server.on('upgrade', function).