Make appIsHealthy function take appId instead of app.
We need to do this because at the moment the websocket receives an
appId as part of the url, and so doesn't have the app to pass to
appIsHealthy. The downside of this is that it means we're querying
the database for the same app every second (since the websocket
app health server calls this appIsHealthy function every second).
A better appraoch is to have the websocket server query the app
once and then just repeatedly pass it to appIsHealthy. We'll make
that refactor soon.