Refactor websocket app logic out of bin/www
Now we have a WebSocket.js file in our lib/ directory. This file
exports a function that can handle upgrading an HTTP request to
WebSockets. So we can now just import this function in our bin/www
and pass it to the server's 'upgrade' event. This function maintains
in its closure references to all the different websocket servers that
were defined in the WebSocket.js file.

I also refactored the upgrade function to be a switch statement to
better reflect the fact that we're expecting to handle multiple
paths.