Make health check web socket obtain its own manager node instance
Before, the health check web socket server called
DockerWrapper.appIsHealthy every second. That method used to first
query the database for the manager node, which meant a database
hit every second for the same object. Now, the web socket server
gets the manager node just once when it first connects to a client
and then passes this to DockerWrapper.appIsHealthy. This means we
hit the database only once per connection.