Add interactive terminal web socket route
This route converts a websocket connection into a full duplex
stream. We then create a plain ubuntu container running bash
and attach it to the websocket stream. This means any data sent
from the client connected to the socket will be sent to the
container's stdin, and any data from the coontainer's stdout
will be streamed to the client.

Note that when we create the container, we ask docker to
allocate a TTY for us. This means that the container will be
able to respond to control sequences from stdin, and that
output from stdout will have the correct ANSI escape sequences
for proper terminal rendering in the client.