Add terminal demo route
This adds a demo route (/terminal) that displays an xtermjs
terminal that connects to our terminal websocket connection, and
allows you to interact with the container via a legit terminal
emulator. It all works -- escape sequences, tab completion, command
history search, interactive commands. Heck, you can even install
and use vim.

This is just a demo, of course. The idea is to use this same
technique from the app show page to allow you to connect to your
app via a console for executing commands like db:migrate and for
debugging. But that's easy now that we have this figured out: just
create a container based on the app's latest image, and then do
what we did here.