Fix layout bug in xtermjs
For some reason, the height of the terminal's viewport was shorter
than the height of the full textarea, so once output got to a certain
height, you couldn't see it on screen. This style change manually
corrects the height, but is obviously very brittle. I'm presuming
that the xtermjs css that we import is interacting badly with bulma
and causing this bug. Haven't gotten to debug yet, though.