Modify terminal websocket to accept command and app query strings
We do this so that on the app show page we can create and attach
to a container for the current app, as well as specify the command
that should be run inside the container (default is bash). We use
query strings because there's not really any other good way of
passing information to our WebSocket handler from the front-end.