Fix update due to ForceUpdate issue
The ForceUpdate flag is just a dummy param so that you can change
something and force the container to update. So we can't set it to
1 every time, instead we need to increment it to force the container
to restart. We need to do this because nothing else about the service
config changed (even the image name is the same, it's just a new image).