Add authentication custom middleware
This simple middleware function determines which routes require authentication,
and redirects unauthenticated users to the login page for routes that do. At the
moment, only the login page can be viewed unauthenticated.

Note that we handle unauthenticated attempts to view api routes slightly
differently than regular routes, because we need to send a JSON response not a
browser redirect response.