Handle password validation in controller
We can't validate in the model that a password is not empty or has a min length,
because we first pass the entered value to bcrypt, which always gives us back a
long, non-empty string. So I don't off the top of my head see a better way of doing
this. Ugh.