Attach logged in user to all requests
This middleware attaches a variable to res.locals representing whether the logged
in users is an admin. By setting it on res.locals, the variable becomes available
for all views (which we need, since every route will render the layout, and the
layout needs to know whether to show the 'Users' link). We're doing this because
handlebars templates won't let us do simple equality checks, so I can't just do
`user.role === 'admin'` from inside the template