Add api login route + logic
When a user tries to login via the api, they must submit a json
body with valid username and password. If these are valid, we
check whether we've already generated a token for this user and
send it as the response. Otherwise, we generate a new token, add
it to the user's tokens field, and send the token as a response.
Now the client can use that token in the authorization header
of future requests to access protected resources.