Add redis for session store
We were previously using express-session's built-in default in-memory
store. This is not production ready, plus it was super annoying
during development because every time the app restarted (i.e.,
every time you made a change), you had to log back in. We could
have also used a file system store, but Redis is the more natural
choice and only very slightly complicates deployment.