Create db backup endpoint for apps
Now /apps/:appId/dbdump creates a SQL dump of the app's database,
sends the sql file as a response, and then deletes the temporary
SQL file. The trick was creating a temporary container based off
the postgres image, and running a pg_dump command from it. We have
the pg_dump command connect to the app's actual database service
(made possible by having our temporary container join the app's
network). We don't specify a file in the pg_dump command, which
results in the command streaming its output to stdout. Dockerode
allows gives us a stream of the container's stdout, so we just
pipe that stream into a file.

When the command finishes running, we remove the temporary container,
send the file as a response, and when that is complete we delete
the temporary SQL file from the server.