Add ability to handle destroying database for app
This fully removes an app's database and all associated cruft (db directory on
paas server, paas db record referencing the db, and the volume the db was using).
It's important to remove all this stuff, especially the volume, so that if the user
later adds a database, it will be clean and empty.

After removing the database service, Docker has an issue where it leaves behind
a dangling volume, even after all containers associated with that service have
stopped and been removed. We can't remove this volume until all containers are
that were using it are gone, and unfortunately there's no good way to no this.
So for now we do the simple thing and just wait five seconds -- I've never seen it
take longer than 2 seconds, plus this timeout isn't set until after we remove the
record from the database, which itself takes some time.