Authenticate websocket requests before upgrading
This adds a `/wsauth` endpoint that we use to authenticate websocket requests
before upgrading them to websocket connections. The general problem we faced
is to handle WebSockets with Node's http.Server, you need to listen for the
`upgrade` event, and then pass a callback that can handle the upgrade to WS
connection.

However, if you listen for and handle this upgrade, then the server never triggers
the normal `request` event. Express, remember, is just a request handler that
only gets called when there is a request event. So in WebSocket upgrade requests,
our Express app never got called. And this was a problem because all of our authentication is done via express middleware.

What to do? After a bunch of bad ideas and dead-ends, the simple solution dawned on me: inside our upgrade handler, we can simply send another request to our server, but one that doesn't contain the 'upgrade' and 'connection' headers that trigger the 'upgrade event. This request will contain the cookie header from
the original request, though, so depending on whether we get back a 200
response, or a 302 redirect to /login, we'll know whether the original WebSocket
request should be considered authenticated.

I brought in the `node-fetch` package to send this extra auth request because
Node's built-in `http.get` and `http.request` methods were way too low-level and
annoying to use.