Change build process to use cloud native buildpack
Instead of Herokuish, we're now using Cloud Native Buildpacks and their pack
utility. The driver we use is one Heroku created and published on Docker Hub
called "heroku/buildpacks". This updates DockerWrapper with the functions we
need for this new process.

First, we have two Dockerfiles. One creates an image based on an image on our
repository (which contains both Docker and pack), but copies in our app and
specifies the pack command to run to build an image for our app. Running that
image creates an image containing the app and everything it needs to run.

However, this image doesn't expose any ports to Docker, and doesn't have specify
a port for the app to listen on. So we have a second Dockerfile that builds from
that image, but exposes a port and sets a PORT environment variable so that our
app exposes and listens on the appropriate port.

So then we just add functions to save both Dockerfiles in the app's directory,
then a function that builds an image from the first Dockerfile, another to actually
run that image (and so generate an image for the app), and finally another func
that builds an image from the second Dockerfile. This last image is what gets used
by Docker service create, and so represents the final app image used in the app's
service.