Fix failing test
We had a test for the case when a user posts to /api/login without
a username and password. The test was expecting a 400 bad request
response, but instead was getting a 500 which crashed the server.
I added a conditional to check for missing username/password and
return early with a 400 response in that case.